+++ genometools/ensembl/extract_protein_coding_exon_annotations.py
@@ -21,10 +21,11 @@
 
 import sys
 import os
-import csv
 import re
 import argparse
 import logging
+
+import unicodecsv as csv
 
 from genometools import misc
 from genometools import ensembl
@@ -42,8 +43,8 @@ def get_argument_parser():
     -----
     This function is used by the `sphinx-argparse` extension for sphinx.
     """
-    description = 'Extract all exon annotations of protein-coding genes.'
-    parser = get_gtf_argument_parser(description, default_field_name = 'exon')
+    desc = 'Extract all exon annotations of protein-coding genes.'
+    parser = get_gtf_argument_parser(desc, default_field_name = 'exon')
     return parser
 
 def main(args=None):
@@ -58,23 +59,19 @@ def main(args=None):
     species = args.species
     chrom_pat = args.chromosome_pattern
     field_name = args.field_name
+
     log_file = args.log_file
     quiet = args.quiet
     verbose = args.verbose
 
-    # configure logger
+    # configure root logger
     log_stream = sys.stdout
     if output_file == '-':
         # if we print output to stdout, redirect log messages to stderr
         log_stream = sys.stderr
 
-    log_level = logging.INFO
-    if quiet:
-        log_level = logging.WARNING
-    elif verbose:
-        log_level = logging.DEBUG
-    logger = misc.configure_logger(__name__, log_stream = log_stream,
-            log_file = log_file, log_level = log_level)
+    logger = misc.get_logger(log_stream = log_stream, log_file = log_file,
+            quiet = quiet, verbose = verbose)
 
     if chrom_pat is None:
         chrom_pat = re.compile(ensembl.species_chrompat[species])
@@ -89,12 +86,12 @@ def main(args=None):
     i = 0
     exons = 0
     logger.info('Parsing data...')
-    with misc.smart_open(input_file,try_gzip=True) as fh, \
-            misc.smart_open_write(output_file,'w') as ofh:
+    with misc.smart_open(input_file, try_gzip = True) as fh, \
+            misc.smart_open_write(output_file) as ofh:
         #if i >= 500000: break
-        reader = csv.reader(fh,dialect='excel-tab')
-        writer = csv.writer(ofh,dialect='excel-tab',lineterminator='\n',
-                quoting=csv.QUOTE_NONE,quotechar='|')
+        reader = csv.reader(fh, dialect = 'excel-tab')
+        writer = csv.writer(ofh, dialect = 'excel-tab', lineterminator = os.linesep,
+                quoting = csv.QUOTE_NONE , quotechar = '|')
         for l in reader:
             i += 1
             #if i % int(1e5) == 0: