+++ genometools/gtf.py
@@ -25,7 +25,7 @@ def parse_attributes(s):
 
     # use regular expression with negative lookbehind to make sure we don't
     # split on escaped semicolons ("\;")
-    attr_sep = re.compile(r"(?<!\\)\s*;\s*")
+    attr_sep = re.compile(ur'(?<!\\)\s*;\s*')
     attr = {}
     atts = attr_sep.split(s)
     for a in atts: